[레거시 코드 리팩터링 - 1단계] 알파(배철환) 미션 제출합니다. (#249)
* docs : 요구사항 정리

* test : domain 계층 테스트 케이스 추가

* test : MenuService에 대한 테스트 코드 추가 및 dto 추가

* test : ProductService에 대한 테스트 코드 추가

* test : MenuGroupService에 대한 테스트 코드 추가

* test : TableService에 대한 테스트 코드 추가

* test : OrderService에 대한 테스트 코드 추가

* test : TableGroupService에 대한 테스트 코드 추가

* refactor : for문을 스트림으로 대체

* refactor : displayname 추가

* refactor : equals & hashcode 삭제

* refactor : MenuGroupServiceTest를 가짜객체로 하도록 수정

* refactor : ProductServiceTest를 가짜객체로 하도록 수정

* refactor : MenuServiceTest를 가짜객체로 하도록 수정

* refactor : OrderServiceTest를 가짜객체로 하도록 수정

* refactor : TableServiceTest를 가짜객체로 하도록 수정

* refactor : TableGroupServiceTest를 가짜객체로 하도록 수정

* refactor : 데이터를 추가하는 코드 추가

step 2 에서 봐요~.

+++ src/main/java/kitchenpos/dao/JdbcTemplateMenuDao.java
@@ -72,7 +72,7 @@ public class JdbcTemplateMenuDao implements MenuDao {
         final Menu entity = new Menu();
         entity.setId(resultSet.getLong("id"));
         entity.setName(resultSet.getString("name"));
-        entity.setPrice(resultSet.getBigDecimal("price"));
+        entity.setPrice(resultSet.getLong("price"));
         entity.setMenuGroupId(resultSet.getLong("menu_group_id"));
         return entity;
     }
